src/agv_transport_web/agv_transport_web/web_dashboard.py
#!/usr/bin/env python3
"""
AGV Transport Web Dashboard Entry Point

This module provides the main entry point for the AGV Transport Web Dashboard.
It will be implemented in subsequent tasks to include FastAPI server setup,
ROS2 integration, and web interface components.
"""

import rclpy
from rclpy.node import Node


class WebDashboardNode(Node):
    """
    ROS2 node for the AGV Transport Web Dashboard.
    
    This is a placeholder implementation that will be expanded
    in subsequent tasks to include:
    - FastAPI server integration
    - ROS2 topic subscriptions and publishers
    - Web interface serving
    """
    
    def __init__(self):
        super().__init__('agv_transport_web_dashboard')
        self.get_logger().info('AGV Transport Web Dashboard node initialized')
        
    def run(self):
        """Run the web dashboard node."""
        self.get_logger().info('Web dashboard node running...')
        # FastAPI server will be started here in subsequent tasks


def main(args=None):
    """Main entry point for the web dashboard."""
    rclpy.init(args=args)
    
    try:
        node = WebDashboardNode()
        node.run()
        rclpy.spin(node)
    except KeyboardInterrupt:
        pass
    finally:
        if rclpy.ok():
            rclpy.shutdown()


if __name__ == '__main__':
    main()