src/agv_transport_web/requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1

# Data validation and serialization
pydantic==2.5.0

# YAML configuration parsing
PyYAML==6.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
hypothesis==6.92.1

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1